Add files via upload
I made a top 10 craft beers of the summer list to go in our BEER section. Here are images of each if we'd like to include them (we don't need to include them all but just a thought!). The first beer on the list is called "Florida Room" and has an image of a gator so we could always include that one as a nod to UF! I can work on the format later but for now, here are the images!